Implemented Dijkstra (70%)

+++ Proj2/Map.h
@@ -8,7 +8,7 @@ struct Map
 
 	~Map()
 	{
-		for (size_t i = 0; i < size.x; i++)
+		for (int i = 0; i < size.x; i++)
 			delete[] tiles[i];
 
 		delete[] tiles;
